renaming ComponentFilesUtils.js to ComponentFilesFactory
